fix(validator): type interval params as float for sub-second test values
CI mypy (`mypy .`) checks the tests directory which the local `mypy
validator/ scoring/` invocation skipped. Loop tests assign sub-second
floats (e.g. 0.01) to challenge_interval_seconds,
weight_check_interval_seconds, and cleanup_interval_seconds so the
loops iterate quickly enough to verify in-process. These attributes
were typed int via their constructor parameters, so mypy rejected the
float assignments across 11 sites in 3 test files.

Widen the three interval parameter annotations to float. They already
represent a duration in seconds that can reasonably be fractional;
the int typing was over-constrained. cleanup_retention_hours stays
int - no test assigns a fraction to it and the DB cleanup API
expects whole hours.

No runtime behaviour change.

+++ validator/main.py
@@ -81,9 +81,9 @@ class ZhenValidator:
         manifest_path: Path = DEFAULT_MANIFEST_PATH,
         boptest_url: str = "http://localhost:8000",
         health_port: int = 8080,
-        challenge_interval_seconds: int = DEFAULT_CHALLENGE_INTERVAL_SECONDS,
-        weight_check_interval_seconds: int = DEFAULT_WEIGHT_CHECK_INTERVAL_SECONDS,
-        cleanup_interval_seconds: int = DEFAULT_CLEANUP_INTERVAL_SECONDS,
+        challenge_interval_seconds: float = DEFAULT_CHALLENGE_INTERVAL_SECONDS,
+        weight_check_interval_seconds: float = DEFAULT_WEIGHT_CHECK_INTERVAL_SECONDS,
+        cleanup_interval_seconds: float = DEFAULT_CLEANUP_INTERVAL_SECONDS,
         cleanup_retention_hours: int = DEFAULT_CLEANUP_RETENTION_HOURS,
     ) -> None:
         """Initialize the validator neuron.
